+++ app/admin/components/vpManagementDirective.js
@@ -87,15 +87,11 @@
                 commonService.getDevelopers()
                     .then(function (developers) {
                         self.developers = developers.developers;
+                        prepCodes();
 
                         if (self.productId) {
                             self.loadCp();
                         }
-                    });
-
-                commonService.getDeveloperTransparencies()
-                    .then(function (developers) {
-                        self.developerMap = developers;
                     });
 
                 getResources();
@@ -136,9 +132,10 @@
                 });
                 self.modalInstance.result.then(function (result) {
                     self.activeDeveloper = result;
-                    commonService.getDevelopers()
+                    commonService.getDeveloperTransparencies()
                         .then(function (developers) {
                             self.developers = developers.developers;
+                            prepCodes();
                         });
                 }, function (result) {
                     if (result !== 'cancelled') {
@@ -162,9 +159,10 @@
                 });
                 self.modalInstance.result.then(function (result) {
                     self.developerMessage = null;
-                    commonService.getDevelopers()
+                    commonService.getDeveloperTransparencies()
                         .then(function (developers) {
                             self.developers = developers.developers;
+                            prepCodes();
                         });
                 }, function (result) {
                     if (result !== 'cancelled') {
@@ -532,6 +530,21 @@
                     .then(function (response) {
                         self.resources.targetedUsers = response;
                     });
+            }
+
+            function prepCodes () {
+                var values = {};
+                for (var i = 0; i < self.developers.length; i++) {
+                    self.developers[i].transMap = {};
+                    for (var j = 0; j < self.developers[i].transparencyAttestations.length; j++) {
+                        self.developers[i].transMap[self.developers[i].transparencyAttestations[j].acbName] = self.developers[i].transparencyAttestations[j].attestation;
+                        values[self.developers[i].transparencyAttestations[j].acbName] = true;
+                    }
+                }
+                self.activeAcbs = [];
+                angular.forEach(values, function (value, key) {
+                    self.activeAcbs.push(key);
+                });
             }
         }]);